+++ CHANGELOG.md
@@ -1,9 +1,14 @@
-### 1.0.0 (2015-05-23)
 
-  * Inital Release
+### 1.2.0 (2015-06-15)
+  * added: request a reports state
+  * added: delete all reports
 
 ### 1.1.0 (2015-05-25)
   * added: logout
   * added: deleteReport
   * added: getApps
   * modified: scheduleReport now supports "all" as appId parameter
+
+### 1.0.0 (2015-05-23)
+
+  * Inital Release